scraper/requirements.txt
scrapy
scrapy-playwright
scrapy-user-agents
scrapy-rotating-proxies
pymongo[srv]
python-dotenv